Extract overlay_lookup_name + reorder_rating_template_vars to output_overlays
Sprint 2, PR #3. Two more nested helpers -> output_overlays.py:
- overlay_lookup_name (6L): canonicalize commonsense overlay aliases.
- reorder_rating_template_vars (60L): stable YAML key order for rating overlays.

The 47-item preferred-order list moves to a module-level tuple
_RATING_TEMPLATE_VAR_ORDER; the commonsense alias set moves to a
module-level frozenset _COMMONSENSE_ALIASES.

Impact:
- modules/output.py: 1311 -> 1247 (-64 / -4.9%)
- modules/output_overlays.py: 321 -> 431 (+110)

Cumulative on output.py: 3833 -> 1247 (-2586 / -67.5%)

All 1081 tests pass.  Ruff + black clean.
