+++ modules/output_overlays.py
@@ -319,3 +319,110 @@ def prune_rating_template_vars(overlay_entry):
         overlay_entry["template_variables"] = cleaned
     else:
         overlay_entry.pop("template_variables", None)
+
+
+_COMMONSENSE_ALIASES = frozenset(
+    {
+        "commonsense",
+        "overlay_content_rating_commonsense",
+        "content_rating_commonsense",
+    }
+)
+
+
+def overlay_lookup_name(name):
+    """Canonicalize legacy commonsense overlay aliases.
+
+    The Quickstart UI, Kometa defaults, and Kometa overlay filenames
+    have disagreed at various times about what to call the
+    commonsense content-rating overlay.  This function collapses the
+    three known aliases into the single canonical
+    ``content_rating_commonsense`` value that Kometa currently
+    expects.  Non-string inputs pass through untouched so callers
+    can hand it any value from the config dict without pre-checks.
+    """
+    if not isinstance(name, str):
+        return name
+    if name in _COMMONSENSE_ALIASES:
+        return "content_rating_commonsense"
+    return name
+
+
+# Preferred YAML emission order for rating-overlay template variables.
+# Keeping this at module scope avoids the ~50-item list rebuild on
+# every call.  Keys not in this list get emitted after in
+# insertion order.
+_RATING_TEMPLATE_VAR_ORDER = (
+    "builder_level",
+    "rating1",
+    "rating1_image",
+    "rating1_font",
+    "rating1_font_size",
+    "rating1_font_color",
+    "rating1_stroke_width",
+    "rating1_stroke_color",
+    "rating1_horizontal_offset",
+    "rating1_vertical_offset",
+    "rating2",
+    "rating2_image",
+    "rating2_font",
+    "rating2_font_size",
+    "rating2_font_color",
+    "rating2_stroke_width",
+    "rating2_stroke_color",
+    "rating2_horizontal_offset",
+    "rating2_vertical_offset",
+    "rating3",
+    "rating3_image",
+    "rating3_font",
+    "rating3_font_size",
+    "rating3_font_color",
+    "rating3_stroke_width",
+    "rating3_stroke_color",
+    "rating3_horizontal_offset",
+    "rating3_vertical_offset",
+    "horizontal_position",
+    "horizontal_offset",
+    "vertical_offset",
+    "flag_alignment",
+    "back_align",
+    "back_color",
+    "back_height",
+    "back_width",
+    "back_line_color",
+    "back_line_width",
+    "back_padding",
+    "back_radius",
+    "use_subtitles",
+)
+
+
+def reorder_rating_template_vars(overlay_entry):
+    """Reorder a rating overlay's ``template_variables`` for stable YAML.
+
+    Mutates ``overlay_entry`` in place.  No-op for non-rating overlays
+    and for entries whose ``template_variables`` is missing, non-dict,
+    or empty.  Keys not in the preferred order are appended in their
+    original insertion order.
+
+    The stable order matters because ruamel.yaml preserves dict
+    insertion order in emitted YAML, and users diff generated configs
+    against previous runs.  Without this reorder, an unpredictable
+    UI-driven insertion order would produce noisy diffs.
+    """
+    if not isinstance(overlay_entry, dict):
+        return
+    default_name = overlay_entry.get("default", "")
+    if not (isinstance(default_name, str) and (default_name == "ratings" or default_name.startswith("overlay_ratings"))):
+        return
+    tv = overlay_entry.get("template_variables")
+    if not isinstance(tv, dict) or not tv:
+        return
+    ordered = {}
+    for key in _RATING_TEMPLATE_VAR_ORDER:
+        if key in tv:
+            ordered[key] = tv[key]
+    for key in tv:
+        if key not in ordered:
+            ordered[key] = tv[key]
+    overlay_entry["template_variables"] = ordered
